feat(utils): updated yup 1.0 usage
