+++ packages/utils/src/index.ts
@@ -1,5 +1,5 @@
 import CookieUtils from "./cookie";
-import yup from "./yup_extended";
+import yup from "./yup_extensions";
 import { makeCancelable, timeout, CancelablePromise } from "./promise";
 
 export { CookieUtils, makeCancelable, timeout, yup };
