+++ packages/utils/src/yup_extensions.ts
@@ -1,0 +1,10 @@
+import { addMethod, string } from "yup";
+
+addMethod(string, "isNumber", function (message?: string) {
+  return this.test("isNumber", message ?? "", (value?: string) => {
+    if (value) {
+      const number = Number(value);
+      return !Number.isNaN(number);
+    }
+  });
+});
